# src/main/resources/static/openapi/admin.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MJU Admin API",
    "version": "v1.0.0",
    "description": "명지대학교 학생회 어드민(ADMIN) 계정 등록 및 관리 기능을 위한 API 명세서입니다."
  },
  "servers": [
    {
      "url": "http://api.mjsearch.org:8080/api/v1"
    }
  ],
  "tags": [
    {
      "name": "Admin",
      "description": "학생회 어드민 계정 등록 및 수정, 검증 기능"
    },
    {
      "name": "DEPARTMENT_NOTICE",
      "description": "학과 공지사항(Department Notice)에 대한 관리자 기능 API입니다. 공지 등록, 수정, 조회, 삭제 등의 기능을 제공합니다."
    }
  ],
  "paths": {
    "/admin/account/register": {
      "post": {
        "tags": ["Admin"],
        "security": [ { "Authorization": [] } ],
        "summary": "초기 관리자 계정 등록",
        "description": "**OPERATOR 전용 API**로, 시스템에 최초 학생회 관리자(ADMIN) 계정을 등록합니다.\n\n\n- 최초 등록된 관리자는 `ADMIN` 권한과 함께 학과 정보를 가지며, 해당 학과의 대표 학생회 계정으로 설정됩니다.\n\n\n- 요청 시, 이메일과 이름, 학과 정보는 필수이며 기본 비밀번호는 서버에서 자동 설정됩니다.\n\n\n- `contactEmail`은 학생회 공식 연락 이메일로 등록되며, 학과 엔티티의 `studentCouncilContactEmail`로 저장됩니다.\n\n\n❗ 이 API는 **운영자(OPERATOR)** 권한을 가진 사용자만 호출할 수 있습니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentCouncilInitRegistrationRequestDTO"
              },
              "example": {
                "email": "mjsearch@mju.ac.kr",
                "name": "데이터사이언스 학생회",
                "contactEmail": "mjsearch2025@gmail.com",
                "departmentName": "DATA_SCIENCE"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "초기 관리자 등록 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                },
                "example": {
                  "status": "API 요청 성공",
                  "data": "초기 관리자가 성공적으로 등록되었습니다.",
                  "timestamp": "2025-07-21T19:55:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "유효성 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": 400,
                  "error": "VALIDATION_FAILED",
                  "message": "email : 올바른 이메일 형식이 아닙니다., contactEmail : 컨텍 이메일은 필수입니다."
                }
              }
            }
          },
          "403": {
            "description": "접근 권한 없음 (OPERATOR 아님)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": 403,
                  "error": "ACCESS_DENIED",
                  "message": "[MJS] 토큰이 유효하지 않거나 사용자에게 해당 접근 권한이 없습니다."
                }
              }
            }
          }
        }
      }
    },
    "/admin/account/update": {
      "patch": {
        "tags": ["Admin"],
        "security": [ { "Authorization": [] } ],
        "summary": "관리자 정보 수정",
        "description": "학생회 관리자 계정과 해당 학과 정보를 동시에 수정합니다. 이름, 학과명, 슬로건, 인스타그램, 공식 홈페이지 URL 등 다양한 속성을 갱신할 수 있습니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentCouncilUpdateDTO"
              },
              "example": {
                "email": "mjsearch@mju.ac.kr",
                "name": "데이터사이언스 학생회",
                "contactEmail": "mjsearch2025@gmail.com",
                "departmentName": "DATA_SCIENCE",
                "password": "newSecurePw1!",
                "college": "AI_SOFTWARE",
                "profileImageUrl": "https://cdn.mjsearch.org/profiles/admin.jpeg",
                "slogan": "함께, 바꾸다!",
                "description": "명지대학교 학생들을 위한 대표 조직입니다.",
                "instagramUrl": "https://instagram.com/mju_student_council",
                "homepageUrl": "https://student.mju.ac.kr"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "관리자 정보 수정 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentCouncilResponseDTO"
                },
                "example": {
                  "status": "API 요청 성공",
                  "data": {
                    "adminEmail": "mjsearch@mju.ac.kr",
                    "name": "데이터사이언스 학생회",
                    "studentCouncilEmail": "mjsearch2025@gmail.com",
                    "college": "AI_SOFTWARE",
                    "departmentName": "DATA_SCIENCE",
                    "profileImageUrl": "https://cdn.mjsearch.org/profiles/admin.jpeg",
                    "slogan": "함께, 바꾸다!",
                    "description": "명지대학교 학생들을 위한 대표 조직입니다.",
                    "instagramUrl": "https://instagram.com/mju_student_council",
                    "homepageUrl": "https://student.mju.ac.kr"
                  },
                  "timestamp": "2025-07-21T19:51:57.6547447"
                }
              }
            }
          }
        }
      }
    },
    "/admin/account/validate": {
      "get": {
        "tags": ["Admin"],
        "security": [ { "Authorization": [] } ],
        "summary": "초기 관리자 계정 여부 확인",
        "description": "이메일을 기준으로 해당 사용자가 시스템에 등록된 초기 관리자(ADMIN)인지 확인합니다. 등록된 경우 true, 없거나 일반 사용자인 경우 false를 반환합니다.",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "description": "확인할 관리자 이메일 주소",
            "schema": {
              "type": "string",
              "format": "email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "검증 결과 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/admin/departments/{departmentUuid}/notices": {
      "post": {
        "tags": ["DEPARTMENT_NOTICE"],
        "security": [ { "Authorization": [] } ],
        "summary": "학과 공지 등록",
        "description": "해당 학과의 학생회 관리자가 새로운 공지사항을 등록합니다. 등록된 공지사항은 학과 학생들에게 노출됩니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminDepartmentNoticeRequestDTO"
              },
              "example": {
                "title": "2025학년도 2학기 개강 안내",
                "content": "안녕하세요. 2학기 개강일은 8월 28일입니다. 수강신청 내역 확인 부탁드립니다.",
                "thumbnailUrl": "https://cdn.mjsearch.org/notice/thumbnail_20250828.png"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "공지 등록 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDepartmentNoticeResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/admin/departments/{departmentUuid}/notices/{noticeUuid}": {
      "get": {
        "tags": ["DEPARTMENT_NOTICE"],
        "security": [ { "Authorization": [] } ],
        "summary": "공지 단건 조회",
        "description": "특정 학과 공지사항의 상세 내용을 조회합니다. 관리 권한이 있는 학생회 어드민만 접근할 수 있습니다.",
        "responses": {
          "200": {
            "description": "공지 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDepartmentNoticeResponseDTO"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["DEPARTMENT_NOTICE"],
        "security": [ { "Authorization": [] } ],
        "summary": "공지 수정",
        "description": "학생회 관리자가 기존 공지사항의 제목, 본문, 썸네일 등을 수정합니다. 비어 있는 필드는 기존 값을 유지합니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminDepartmentNoticeRequestDTO"
              },
              "example": {
                "title": "2학기 개강일 안내 (수정)",
                "content": "개강일은 8월 28일이며, 첫 수업은 온라인으로 진행됩니다.",
                "thumbnailUrl": "https://cdn.mjsearch.org/notice/updated_thumbnail.png"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "공지 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDepartmentNoticeResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["DEPARTMENT_NOTICE"],
        "security": [ { "Authorization": [] } ],
        "summary": "공지 삭제",
        "description": "학생회 관리자가 특정 공지사항을 삭제합니다. 삭제 시 관련된 S3 리소스도 정리됩니다.",
        "responses": {
          "200": {
            "description": "삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string",
                      "example": "삭제가 성공적으로 완료되었습니다."
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "Authorization": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "StudentCouncilInitRegistrationRequestDTO": {
        "type": "object",
        "required": ["email", "name", "contactEmail", "departmentName"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string" },
          "contactEmail": { "type": "string", "format": "email" },
          "departmentName": { "type": "string" }
        }
      },
      "StudentCouncilUpdateDTO": {
        "type": "object",
        "required": ["email", "name", "password", "departmentName", "college"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string" },
          "password": { "type": "string" },
          "college": { "type": "string" },
          "departmentName": { "type": "string" },
          "profileImageUrl": { "type": "string" },
          "slogan": { "type": "string" },
          "description": { "type": "string" },
          "instagramUrl": { "type": "string" },
          "homepageUrl": { "type": "string" },
          "contactEmail": { "type": "string" }
        }
      },
      "StudentCouncilResponseDTO": {
        "type": "object",
        "properties": {
          "adminEmail": { "type": "string" },
          "name": { "type": "string" },
          "studentCouncilEmail": { "type": "string" },
          "college": { "type": "string" },
          "departmentName": { "type": "string" },
          "profileImageUrl": { "type": "string" },
          "slogan": { "type": "string" },
          "description": { "type": "string" },
          "instagramUrl": { "type": "string" },
          "homepageUrl": { "type": "string" }
        }
      },
      "ApiResponseString": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "API 요청 성공" },
          "data": { "type": "string", "example": "초기 어드민 계정이 성공적으로 등록되었습니다." },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
      "ApiResponseStudentCouncilResponseDTO": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "API 요청 성공" },
          "data": { "$ref": "#/components/schemas/StudentCouncilResponseDTO" },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
      "ApiResponseBoolean": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "API 요청 성공" },
          "data": { "type": "boolean", "example": true },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
      "AdminDepartmentNoticeRequestDTO": {
        "type": "object",
        "description": "관리자가 공지사항을 등록하거나 수정할 때 사용하는 요청 DTO",
        "properties": {
          "title": {
            "type": "string",
            "description": "공지사항 제목 (필수)",
            "example": "2025학년도 2학기 개강 안내"
          },
          "content": {
            "type": "string",
            "description": "공지사항 본문 내용 (HTML 가능)",
            "example": "안녕하세요. 2학기 개강일은 8월 28일입니다. 수강신청 내역 확인 부탁드립니다."
          },
          "thumbnailUrl": {
            "type": "string",
            "description": "공지사항 썸네일 이미지의 S3 URL (선택)",
            "example": "https://cdn.mjsearch.org/notice/thumbnail_20250828.png"
          }
        },
        "required": ["title"]
      },
      "AdminDepartmentNoticeResponseDTO": {
        "type": "object",
        "description": "공지 단건 또는 등록/수정 후 반환되는 응답 DTO",
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid",
            "description": "공지사항 고유 식별자",
            "example": "d3cbb3a4-37a3-11ee-be56-0242ac120002"
          },
          "title": {
            "type": "string",
            "description": "공지 제목",
            "example": "2025학년도 2학기 개강 안내"
          },
          "content": {
            "type": "string",
            "description": "공지 본문",
            "example": "2학기 개강일은 8월 28일이며, 수강신청 후 첫 수업은 온라인으로 진행됩니다."
          },
          "createAt": {
            "type": "string",
            "format": "date-time",
            "description": "공지 작성 일시 (ISO 8601 포맷)",
            "example": "2025-07-20T15:00:00"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 400,
            "description": "HTTP 상태 코드"
          },
          "error": {
            "type": "string",
            "example": "VALIDATION_FAILED",
            "description": "에러 식별용 키워드"
          },
          "message": {
            "type": "string",
            "example": "email : 올바른 이메일 형식이 아닙니다., contactEmail : 컨텍 이메일은 필수입니다.",
            "description": "에러 상세 메시지"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-21T19:55:00.000Z",
            "description": "에러 발생 시간 (ISO 8601 형식)"
          }
        },
        "required": ["status", "error", "message", "timestamp"]
      }
    }
  }
}